Add taxonomies and archetype for posts. All pieces will
conform to the post archetype at this time. We intend to migrate all
pieces from the existing site as simply as possible. The Wordpress
site stores all pieces as posts. We follow suit.

Front matter:

However we chose to enrich front matter so that we may automatically
group material in various dimensions, such as Author, Genre, Form or
type (category), Subject matter (tags).

Chronological archives:

Also we may want to expose lists of content by year and/or months. We
have to add a clever workaround for such chronological archive lists
because Hugo does not do it automatically at this time. Ref:
https://discourse.gohugo.io/t/how-to-generate-chronological-blog-archives-in-hugo/13491/6