Add Open Font Licensed fonts for masthead and body
Oswald is based on the classic Gothic typeface tradition. I feel it sits
well with the idea of Bombay being a mix of the old and the new, and of
being a place of stark contrasts. A good story is a stark contrast from
the here and now because it takes us to someplace else entirely.

Eczar is a lovely face from Rosetta Type Foundry, with support for Latin
and Devanagri scripts. But it may have too much personality. Perhaps we
can choose a classic face that does not colour the text too much.